modularize BoardAppShareSubmenu
